Load OBJ

+++ include/renderer.hpp
@@ -32,7 +32,7 @@ public:
     void UpdateCamera();
 
 private:
-    uint16_t m_IndexCount;
+    uint32_t m_IndexCount;
     GLuint m_VAO, m_VBO, m_IBO, m_UBO, m_ShaderProgram;
 
     glm::mat4* m_UBOData = nullptr;
